backend/requirements.txt
Django==5.0.2
djangorestframework
djangorestframework-simplejwt
djoser

psycopg2-binary
django-cors-headers

celery
redis

channels
channels-redis

llama-index
llama-index-embeddings-openai
llama-index-vector-stores-chroma

chromadb

openai

pypdf

python-dotenv

gunicorn
uvicorn
